style: size_t

+++ vehicle/raw_vehicle_cmd_converter/src/csv_loader.cpp
@@ -91,7 +91,7 @@ std::vector<double> CSVLoader::getRowIndex(const Table & table)
 std::vector<double> CSVLoader::getColumnIndex(const Table & table)
 {
   std::vector<double> index = {};
-  for (unsigned int i = 1; i < table.size(); i++) {
+  for (size_t i = 1; i < table.size(); i++) {
     index.push_back(std::stod(table[i][0]));
   }
   return index;
